+++ backend/de.metas.manufacturing/src/test/java/de/metas/material/planning/pporder/impl/PPOrderBOMBLTest.java
@@ -14,12 +14,13 @@ import org.adempiere.model.InterfaceWrapperHelper;
 import org.compiere.SpringContextHolder;
 import org.compiere.model.I_C_UOM;
 import org.compiere.model.I_M_Product;
+import org.eevolution.api.impl.ProductBOMVersionsDAO;
 import org.eevolution.model.I_PP_Order;
 import org.eevolution.model.I_PP_Order_BOMLine;
 import org.eevolution.model.I_PP_Product_BOM;
+import org.eevolution.model.I_PP_Product_BOMVersions;
 import org.eevolution.mrp.api.impl.MRPTestHelper;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
@@ -32,7 +33,6 @@ import static org.mockito.Mockito.mock;
  *
  * @author metas-dev <dev@metasfresh.com>
  */
-@Disabled
 public class PPOrderBOMBLTest
 {
 	private MRPTestHelper helper;
@@ -52,6 +52,7 @@ public class PPOrderBOMBLTest
 		helper = new MRPTestHelper();
 
 		SpringContextHolder.registerJUnitBean(new EventLogService(mock(EventLogsRepository.class)));
+		SpringContextHolder.registerJUnitBean(new ProductBOMVersionsDAO());
 
 		ppOrderBOMDAO = Services.get(IPPOrderBOMDAO.class);
 		ppOrderBOMBL = (PPOrderBOMBL)Services.get(IPPOrderBOMBL.class);
@@ -97,11 +98,13 @@ public class PPOrderBOMBLTest
 				.fromToMultiplier(new BigDecimal("1500000"))
 				.build());
 
+		final I_PP_Product_BOMVersions bomVersions = helper.createBOMVersions(ProductId.ofRepoId(pSalad.getM_Product_ID()));
 		//
 		// Define BOM
 		//@formatter:off
 		final I_PP_Product_BOM saladProductBom = helper.newProductBOM()
 				.product(pSalad).uom(uomStuck)
+				.bomVersions(bomVersions)
 				// Carrot
 				.newBOMLine()
 					.product(pCarrot).uom(uomKillogram)